more WIP for msys2 builds on windows

+++ src/lib/platform/MSWindowsClipboardBitmapConverter.cpp
@@ -2,11 +2,11 @@
  * synergy -- mouse and keyboard sharing utility
  * Copyright (C) 2012-2016 Symless Ltd.
  * Copyright (C) 2004 Chris Schoeneman
- * 
+ *
  * This package is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * found in the file LICENSE that should have accompanied this file.
- * 
+ *
  * This package is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -125,7 +125,7 @@ MSWindowsClipboardBitmapConverter::toIClipboard(HANDLE data) const
 	}
 	else {
 		//http://msdn.microsoft.com/en-us/library/ke55d167(VS.80).aspx
-		srcBits += (1i64 << bitmap->bmiHeader.biBitCount) * sizeof(RGBQUAD);
+		srcBits += (((long long)1) << bitmap->bmiHeader.biBitCount) * sizeof(RGBQUAD);
 	}
 
 	// copy source image to destination image
